#640 [refactor] Resolve merge conflict and apply Long type

+++ presentation/src/main/java/daily/dayo/presentation/viewmodel/FolderViewModel.kt
@@ -88,7 +88,7 @@ class FolderViewModel @Inject constructor(
         }
     }
 
-    fun toggleFolderOrder(folderId: Int) {
+    fun toggleFolderOrder(folderId: Long) {
         val newOrder = when (_uiState.value.folderOrder) {
             FolderOrder.NEW -> FolderOrder.OLD
             FolderOrder.OLD -> FolderOrder.NEW
